pyproject.toml
[project]
name = "ai-jherkin-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "fastapi>=0.121.2",
    "google-generativeai>=0.8.5",
    "lxml>=6.0.2",
    "playwright>=1.56.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.38.0",
]

[tool.uv.scripts]
dev = "uvicorn main:app --reload"
